move binding model into lib/, move data into data
