+++ data/copt_lemma_lex_cplx_2.8_cdo.tab
