--- eval/binding/const.py
@@ -1,1 +1,0 @@
-UNKNOWN_CHAR = "❤"
