+++ lib/binding/const.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+UNKNOWN_CHAR = "❤"
+IGNORE = [
+	"",  #	 empty string
+	"~",  # U+007E  TILDE
+	"`",  # U+0060  GRAVE ACCENT
+	"\n",  #	NEWLINE
+	"̈",  # U+0308  COMBINING DIAERESIS
+	"̄",  # U+0304  COMBINING MACRON
+	"̀",  # U+0300  COMBINING GRAVE ACCENT
+	"̣",  # U+0323  COMBINING DOT BELOW
+	"̅",  # U+0305  COMBINING OVERLINE
+	"̂",  # U+0302  COMBINING CIRCUMFLEX ACCENT
+	"︤",  # U+FE24  COMBINING MACRON LEFT HALF
+	"︥",  # U+FE25  COMBINING MACRON RIGHT HALF
+	"︦",  # U+FE26  COMBINING CONJOINING MACRON
+	"̇",  # U+0307  COMBINING DOT ABOVE
+	"᷍",  # U+1DCD  COMBINING DOUBLE CIRCUMFLEX ABOVE
+	"⳿",  # U+2CFF  COPTIC MORPHOLOGICAL DIVIDER
+]
+GOLD_TOKEN_SEPARATOR = "_"
+ORIG_TOKEN_SEPARATOR = " "
